Migrate to EUPL as the default copyleft-ish license
This resolves some concerns that a few people voiced about "custom" licenses - which is not really how amendments to the GPL work, but people should not be expected to understand all of the legal nuances - while preserving my desired license traits of network loophole closing with minimal risk of being taken away from the public.

There have been a lot of criticisms of the GPL over the years, the Affero branch remedied some of them, but didn't provide a "lesser" variant and continued to rely on copyright for enforcement rather than standard contract terms.

The state of the FSF over the last several years has completely soured me on being affiliated with them in even the most tangential way.

This change resolves all of those issues as well.